Move Compare and CompareWith to IGitService
IGitService knows about IRepository but not remote servers. IGitClient
knows about remote git repositories. IGitService  is a more appropriate
home for Compare and CompareWith.
