+++ src/GitHub.App/SampleData/GitServiceDesigner.cs
@@ -15,5 +15,7 @@ namespace GitHub.SampleData
         public IRepository GetRepository(string path) => null;
         public UriString GetUri(string path, string remote = "origin") => null;
         public UriString GetUri(IRepository repository, string remote = "origin") => null;
+        public Task<Patch> Compare(IRepository repository, string sha1, string sha2, string path) => null;
+        public Task<ContentChanges> CompareWith(IRepository repository, string sha1, string sha2, string path, byte[] contents) => null;
     }
 }
