+++ src/GitHub.Exports/Services/IGitService.cs
@@ -71,5 +71,30 @@ namespace GitHub.Services
         /// <param name="remote">The remote name to look for</param>
         /// <returns></returns>
         Task<string> GetLatestPushedSha(string path, string remote = "origin");
+
+        /// <summary>
+        /// Compares a file in two commits.
+        /// </summary>
+        /// <param name="repository">The repository</param>
+        /// <param name="sha1">The SHA of the first commit.</param>
+        /// <param name="sha2">The SHA of the second commit.</param>
+        /// <param name="path">The relative path to the file.</param>
+        /// <returns>
+        /// A <see cref="Patch"/> object or null if one of the commits could not be found in the repository.
+        /// </returns>
+        Task<Patch> Compare(IRepository repository, string sha1, string sha2, string path);
+
+        /// <summary>
+        /// Compares a file in a commit to a string.
+        /// </summary>
+        /// <param name="repository">The repository</param>
+        /// <param name="sha1">The SHA of the first commit.</param>
+        /// <param name="sha2">The SHA of the second commit.</param>
+        /// <param name="path">The relative path to the file.</param>
+        /// <param name="contents">The contents to compare with the file.</param>
+        /// <returns>
+        /// A <see cref="Patch"/> object or null if the commit could not be found in the repository.
+        /// </returns>
+        Task<ContentChanges> CompareWith(IRepository repository, string sha1, string sha2, string path, byte[] contents);
     }
 }
